Deprecate unused metrics db stuff, cli argument stuff, in data import.
