+++ spatialprofilingtoolbox/workflow/tabular_import/integrator.py
@@ -1,22 +1,12 @@
 """The wrap-up task of the main data import workflow."""
 
-from spatialprofilingtoolbox.workflow.tabular_import.computational_design import TabularImportDesign
-from spatialprofilingtoolbox.standalone_utilities.log_formats import colorized_logger
 
-logger = colorized_logger(__name__)
-
-
-class TabularImportIntegrator:
+class TabularImportIntegrator: #pylint: disable=too-few-public-methods
     """
     Wrap-up phase of the data import workflow. Currently no wrap-up is needed.
     """
     def __init__(self, **kwargs): # pylint: disable=unused-argument
-        self.computational_design = TabularImportDesign(**kwargs)
+        pass
 
-    def get_computational_design(self):
-        return self.computational_design
-
-    def calculate(self, filename):
-        logger.info('Doing integration.')
-        with open(filename, 'wt', encoding='utf-8') as file:
-            file.write('')
+    def calculate(self):
+        pass
